requirements.txt
flask
google-generativeai
python-dotenv
langchain-google-genai
langchain
langchain-community
faiss-cpu
unstructured
supabase-py
gunicorn